Реализован ресайз
