+++ js/screens/game3-view.js
@@ -1,14 +1,12 @@
 import AbstractView from "../abstract-view";
-import {resizeImages} from '../service/image-handler';
 
 export default class GameThirdView extends AbstractView {
-  constructor(state, level, questions, imgs, answers) {
+  constructor(state, level, questions, answers) {
     super();
     this.state = state;
     this.level = level;
     this.answers = answers;
     this.questions = questions;
-    this.imgs = imgs;
   }
 
   get template() {
@@ -40,40 +38,19 @@ export default class GameThirdView extends AbstractView {
   }
 
   bind() {
-    const imgs = this.element.querySelectorAll(`img`);
-    imgs.forEach((item) => item.addEventListener(`load`, () => {
-      const blockWidth = this.level.answers[0].image.width;
-      const blockHeight = this.level.answers[0].image.height;
-      const blockData = {width: blockWidth, height: blockHeight};
-
-      const imgSizes = {width: item.naturalWidth, height: item.naturalHeight};
-      let data = this.resizeImages(blockData, imgSizes);
-      item.setAttribute(`height`, data.height);
-      item.setAttribute(`width`, data.width);
-    }));
-
     const gameForm = this.element.querySelector(`.game__content`);
     gameForm.addEventListener(`click`, (evt) => {
       // Setting variables
       const levelAnswers = this.questions[this.state.level].answers;
       const levelImages = evt.target.children;
-      let mistake;
-
       // Checking if a user ckicked correct div item
       if (evt.target.tagName !== `DIV` && !evt.target.classList.contains(`game__option`)) {
         return;
       }
-
       // Looping through answers set to check which src equals to the chosen src
       const chosenElement = levelAnswers.filter((item) => levelImages[0].src === item.image.url);
-
       // If chosen element equals to the right answer
-      if (chosenElement[0].type !== `paint`) {
-        mistake = true;
-      } else {
-        mistake = false;
-      }
-
+      const mistake = chosenElement[0].type !== `paint`;
       this.onAnswer(mistake);
     });
   }
